+++ apps/platform/public/svg/shared/index.ts
@@ -37,6 +37,7 @@ import InvitationSVG from './invitation.svg'
 import ErrorInfoSVG from './errorInfo.svg'
 import DiscordSVG from './discord.svg'
 import SlackSVG from './slack.svg'
+import VercelSVG from './vercel.svg'
 import EmptyIntegrationSVG from './emptyIntegration.svg'
 import PendingSVG from './pending.svg'
 
@@ -81,5 +82,6 @@ export {
   ErrorInfoSVG,
   DiscordSVG,
   SlackSVG,
-  PendingSVG
+  PendingSVG,
+  VercelSVG
 }