+++ packages/schema/src/enums/index.ts
@@ -4,7 +4,8 @@ export const integrationTypeEnum = z.enum([
   'DISCORD',
   'SLACK',
   'GITHUB',
-  'GITLAB'
+  'GITLAB',
+  'VERCEL'
 ])
 
 export const expiresAfterEnum = z.enum(['never', '24', '168', '720', '8760'])